(PSP) fix build.

+++ dist-scripts/psp1-cores.sh
@@ -16,16 +16,27 @@ for f in *_psp1.a ; do
    name=`echo "$f" | sed 's/\(_libretro_psp1\|\).a$//'`
    whole_archive=
    big_stack=
+
    if [ $name = "nxengine" ] ; then
       echo "NXEngine found, applying whole archive linking..."
       whole_archive="WHOLE_ARCHIVE_LINK=1"
    fi
+
    if [ $name = "tyrquake" ] ; then
       echo "Tyrquake found, applying big stack..."
       big_stack="BIG_STACK=1"
    fi
+
+   if [ $big_stack="BIG_STACK=1" ] ; then
+      make -C ../ -f Makefile.psp1 clean || exit 1
+   fi
+
    cp -f "$f" ../libretro_psp1.a
    make -C ../ -f Makefile.psp1 $whole_archive $big_stack -j3 || exit 1
    mv -f ../EBOOT.PBP ../psp1/pkg/cores/${name}_libretro.PBP
    rm -f ../retroarchpsp.elf
+
+   if [ $big_stack="BIG_STACK=1" ] ; then
+      make -C ../ -f Makefile.psp1 clean || exit 1
+   fi
 done
